refactor libretto-cli to yargs modules and add follow-up spec
