+++ packages/libretto-cli/package.json
@@ -16,10 +16,12 @@
 	"dependencies": {
 		"libretto": "workspace:*",
 		"playwright": "^1.52.0",
+		"yargs": "^17.7.2",
 		"zod": "^3.25.0"
 	},
 	"devDependencies": {
 		"@types/node": "^25.3.3",
+		"@types/yargs": "^17.0.33",
 		"tsup": "^8.0.0",
 		"tsx": "^4.19.2",
 		"typescript": "^5.7.0"
